perf: defer dashboard degen graph
